pyproject.toml
[tool.poetry]
name = "securestep-rag"
version = "0.1.0"
description = "Hardened iterative RAG pipeline against adversarial prompt injection attacks"
authors = ["SecureStep Team <team@example.com>"]
readme = "README.md"
packages = [
  { include = "pipeline" },
  { include = "trust_filter" },
  { include = "attack" },
  { include = "models" },
  { include = "vector_store" },
  { include = "eval" },
  { include = "benchmark" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain-core = "^0.3.0"
langgraph = "^0.2.30"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.3.0"
langchain-community = "^0.3.0"
llama-index = "^0.11.0"
qdrant-client = "^1.12.0"
FlagEmbedding = "^1.3.2"
ragas = "^0.2.5"
deepeval = "^1.4.0"
wandb = "^0.18.0"
arize-phoenix = "^6.0.0"
nemoguardrails = "^0.11.0"
garak = "^0.10.0"
pyrit = "^0.5.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
redis = "^5.1.1"
celery = "^5.4.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
datasets = "^3.1.0"
huggingface-hub = "^0.26.0"
numpy = "^2.1.0"
scikit-learn = "^1.5.2"
pydantic = "^2.9.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
ruff = "^0.7.1"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
